+++ vec/src/Data/Vec/Lazy.hs
@@ -37,6 +37,8 @@ module Data.Vec.Lazy (
     snoc,
     head,
     tail,
+    -- * Reverse
+    reverse,
     -- * Concatenation and splitting
     (++),
     split,
@@ -435,7 +437,9 @@ cons :: a -> Vec n a -> Vec ('S n) a
 cons = (:::)
 
 -- | Add a single element at the end of a 'Vec'.
-snoc ::Vec n a -> a -> Vec ('S n) a
+--
+-- @since 0.2.1
+snoc :: Vec n a -> a -> Vec ('S n) a
 snoc VNil       x = x ::: VNil
 snoc (y ::: ys) x = y ::: snoc ys x
 
@@ -446,6 +450,21 @@ head (x ::: _) = x
 -- | The elements after the 'head' of a 'Vec'.
 tail :: Vec ('S n) a -> Vec n a
 tail (_ ::: xs) = xs
+
+-------------------------------------------------------------------------------
+-- Reverse
+-------------------------------------------------------------------------------
+
+-- | Reverse 'Vec'.
+--
+-- >>> reverse ('a' ::: 'b' ::: 'c' ::: VNil)
+-- 'c' ::: 'b' ::: 'a' ::: VNil
+--
+-- @since 0.2.1
+--
+reverse :: Vec n a -> Vec n a
+reverse VNil       = VNil
+reverse (x ::: xs) = snoc (reverse xs) x
 
 -------------------------------------------------------------------------------
 -- Concatenation