+++ vec/src/Data/Vec/Pull.hs
@@ -32,6 +32,8 @@ module Data.Vec.Pull (
     snoc,
     head,
     tail,
+    -- * Reverse
+    reverse,
     -- * Folds
     foldMap,
     foldMap1,
@@ -284,6 +286,8 @@ cons x (Vec v) = Vec $ \i -> case i of
     FS j -> v j
 
 -- | Add a single element at the end of a 'Vec'.
+--
+-- @since 0.2.1
 snoc :: forall a n. N.InlineInduction n => Vec n a -> a -> Vec ('S n) a
 snoc (Vec xs) x = Vec $ \i -> case F.isMax i of
     Nothing -> x
@@ -296,6 +300,20 @@ head (Vec v) = v FZ
 -- | The elements after the 'head' of a 'Vec'.
 tail :: Vec ('S n) a -> Vec n a
 tail (Vec v) = Vec (v . FS)
+
+-------------------------------------------------------------------------------
+-- Reverse
+-------------------------------------------------------------------------------
+
+-- | Reverse 'Vec'.
+--
+-- >>> reverse ('a' ::: 'b' ::: 'c' ::: VNil)
+-- 'c' ::: 'b' ::: 'a' ::: VNil
+--
+-- @since 0.2.1
+--
+reverse :: forall n a. N.InlineInduction n => Vec n a -> Vec n a
+reverse (Vec v) = Vec (v . F.mirror)
 
 -------------------------------------------------------------------------------
 -- Mapping