+++ vec/test/Inspection/DataFamily/SpineStrict.hs
@@ -81,3 +81,15 @@ rhsSnoc :: Vec N.Nat3 Char
 rhsSnoc = 'a' ::: 'b' ::: 'c' ::: VNil
 
 inspect $ 'lhsSnoc  === 'rhsSnoc
+
+-------------------------------------------------------------------------------
+-- reverse
+-------------------------------------------------------------------------------
+
+lhsReverse :: Vec N.Nat3 Char
+lhsReverse = I.reverse $ 'c' ::: 'b' ::: 'a' ::: VNil
+
+rhsReverse :: Vec N.Nat3 Char
+rhsReverse = 'a' ::: 'b' ::: 'c' ::: VNil
+
+inspect $ 'lhsReverse  === 'rhsReverse